third_party/blink/web_tests/wpt_internal/prerender/restriction-pointer-lock.html

<!DOCTYPE html>
<!-- This cannot be upstreamed as a WPT test because the spec (probably) will
     require that no error is thrown until activation. -->
<title>Same-origin prerendering cannot request PointerLock</title>
<meta name="timeout" content="long">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/common/utils.js"></script>
<script src="/speculation-rules/prerender/resources/utils.js"></script>
<body>
<script>

promise_test(async t => {
  const uid = token();
  const bc = new PrerenderChannel('prerender-channel', uid);

  const gotMessage = new Promise(resolve => {
    bc.addEventListener('message', e => {
      resolve(e.data);
    }, {
      once: true
    });
  });

  // Start prerendering a page that attempts to request pointer lock.
  startPrerendering(`resources/pointer-lock.html?uid=${uid}`);
  const result = await gotMessage;
  assert_equals(result, 'WrongDocumentError');
}, 'prerendering pages should not be able to request pointer lock');

</script>
</body>
